requirements.txt
# --- Core ---
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.23.0

# --- LLM & Vector DB ---
groq>=0.5.0
qdrant-client>=1.1.0
fastembed>=0.2.0

# --- Data Processing ---
pypdf>=4.0.0
numpy
tqdm

# --- Optional / Legacy ---
# sentence-transformers>=2.2.0  # Optional if using local heavy models
